requirements.txt
concrete-python
numba
numpy
spacy
nltk
tqdm
keras
pandas
codecs
sklearn
matplotlib
